+++ package/MDAnalysis/converters/accessors.py
